feat: [DP-1690] - getting users data on a history page
